Se agrego API Ordenes Produccion

+++ Panaderia/Models/Detalles_de_compras.cs
@@ -10,9 +10,9 @@ namespace Panaderia.Models
         public int fk_compra { get; set; }
         public int fk_ingrediente { get; set; }
         public int fk_stock { get; set; }
-        public decimal dc_precio_unidad { get; set; }
+        public float dc_precio_unidad { get; set; }
         public int int_cantidad { get; set; }
-        public decimal int_iva { get; set; }
+        public int int_iva { get; set; }
 
         public virtual Compras? Compras { get; set; }
         public virtual Ingredientes? Ingredientes { get; set; }
